fix delte post

+++ components/Post.js
@@ -7,7 +7,7 @@ import {HeartIcon as HeartIconFilled} from "@heroicons/react/solid"
 import {signIn, useSession} from "next-auth/react";
 import { deleteObject, ref } from 'firebase/storage'
 
-function Post({post}) {
+function Post({post,id}) {
   const {data:session} = useSession();
   const [likes,setLikes] = useState([]);
   const [hasLiked,setHasLiked] = useState(false);
@@ -20,7 +20,7 @@ function Post({post}) {
   }, [db]);
 
   useEffect(()=>{
-   setHasLiked(likes.findIndex((like)=>like.id === session?.user.uid) !== -1);
+   setHasLiked(likes.findIndex((like) => like.id === session?.user.uid) !== -1);
   }, [likes])
 
   async function likePost(){
@@ -38,27 +38,33 @@ function Post({post}) {
   }
 
   async function deletePost() {
-    if(window.confirm('Are you sure you want to delete the post?')){
-      deleteDoc(doc(db, "posts", post.id))
-      deleteObject(ref(storage,`posts/${post.id}/image`))
+    if(window.confirm("Are you sure you want to delete the post?")){
+      deleteDoc(doc(db, "posts", post.id));
+      if(post.data().image){
+        deleteObject(ref(storage,`posts/${post.id}/image`));
+      }
+
     }
   }
 
   return (
     <div className="flex p-3 cursor-pointer border-b border-gray-200">
         {/* user image */}
-        <img className="h-11 w-11 rounded-full mr-4"src={post.data().userImg} alt="user-img"/>
+        <img className="h-11 w-11 rounded-full mr-4"
+        src={post?.data()?.userImg}
+        alt="user-img"
+      />
         {/* right side */}
-        <div className="">
+        <div className="flex-1">
             {/* header */}
 
             <div className="flex items-center justify-between">
                 {/* user info*/}
                 <div className="flex items-center space-x-1 whitespace-nowrap">
-                <h4 className="font-bold text-[15px] sm:text-[16px] hover:underline">{post.data().name}</h4>
-                <span className="text-sm sm:text-[15px]">@{post.data().username} - </span>
+                <h4 className="font-bold text-[15px] sm:text-[16px] hover:underline">{post?.data()?.name}</h4>
+                <span className="text-sm sm:text-[15px]">@{post?.data()?.username} - </span>
                 <span className="text-sm sm:text-[15px] hover:underline">
-                <Moment fromNow>{post?.data().timestamp?.toDate()}</Moment>
+                <Moment fromNow>{post?.data()?.timestamp?.toDate()}</Moment>
                 </span>
                 </div>
                 {/* dot icon */}
@@ -66,14 +72,14 @@ function Post({post}) {
             </div>
         
           {/* post text */}
-           <p className="text-gray-800 text-[15px] sm:text-[16px] mb-2">{post.data().text}</p>
+           <p className="text-gray-800 text-[15px] sm:text-[16px] mb-2">{post?.data()?.text}</p>
           {/* post image */}
-          <img className="rounded-2xl mr-2"src={post.data().image} alt=""/>
+          <img className="rounded-2xl mr-2"src={post?.data()?.image} alt=""/>
           {/* icons */}
 
           <div className="flex justify-between text-gray-500 p-2">
             <ChatIcon className="h-9 w-9 hoverEffect p-2 hover:text-sky-500 hover:bg-sky-100"/>
-            {session?.user.uid === post?.data().id && (
+            {session?.user.uid === post?.data()?.id && (
             <TrashIcon onClick={deletePost}className="h-9 w-9 hoverEffect p-2 hover:text-red-600 hover:bg-red-100"/>
             )}
